+++ CHANGES.rst
@@ -20,6 +20,7 @@ in progress
 - Compensate empty values (nan) when importing from CSV
 - Add output formatter for Grafana Worldmap Panel JSON file
 - Add RDBMS database (PostgreSQL) as station data source
+- Add ``--sensor-type`` filter option
 
 
 2018-12-11 0.4.3